dotenv
Going to support the two ways of receiving the port number. First, we're going to give priority to an environment variable. So if we set it via an environment variable, that's the one that is going to be used to start up the server. If there is no environment variable available, then we are going to check a command line argument. And finally, if neither of those are present, then we're going to go for a default port number.